add initial support for leancrypto
By using the leancrypto library, the shim boot loader receives full
support for (accelerated) PQC and composite PQC algorithms.

The patch adds the leancrypto support in parallel to the OpenSSL support
to ensure that the old algorithms of SHA-1 and RSA are still present and
usable in parallel with the leancrypto algorithms.

The README.leancrypto.md contains information about the compilation and
usage of leancrypto with Shim.

Note: The composite PQC support specified for X.509 in [1] is currently
work in progress. This implies that the X.509 structure may be subject
to change which will be integrated into leancrypto. Therefore, consider
the composite PQC support as not yet stable.

Note2: This patch is work in progress. This patch is only compile-tested
as of now.

Note3: The renaming of the OpenSSL Cryptlib interface functions has been
made as it is currently not clear that they are used at all. My currently
renaming them, the linker shall tell us if they need to be converted to
leancrypto.

[1] https://lamps-wg.github.io/draft-composite-sigs/draft-ietf-lamps-pq-composite-sigs.html

Signed-off-by: Stephan Mueller <smueller@chronox.de>